(PC-21481) chore(types): extract new Offer type from algolia folder into a shared type file
