+++ src/features/search/components/SearchListFooter/SearchListFooter.web.tsx
@@ -2,7 +2,7 @@ import React, { forwardRef } from 'react'
 import { ActivityIndicator } from 'react-native'
 import styled from 'styled-components/native'
 
-import { Offer } from 'libs/algolia'
+import { Offer } from 'shared/offer/types'
 import { ButtonSecondary } from 'ui/components/buttons/ButtonSecondary'
 import { More } from 'ui/svg/icons/More'
 import { getSpacing, Spacer } from 'ui/theme'
