+++ src/libs/algolia/__mocks__/mockedAlgoliaResponse.ts
@@ -1,7 +1,7 @@
 import { SearchResponse } from '@algolia/client-search'
 
 import { SubcategoryIdEnum } from 'api/gen'
-import { Offer } from 'libs/algolia'
+import { Offer } from 'shared/offer/types'
 
 export const mockedAlgoliaResponse: SearchResponse<Offer> = {
   hits: [
