+++ src/libs/algolia/fetchAlgolia/fetchMultipleOffers/fetchMultipleOffers.ts
@@ -1,7 +1,6 @@
 import flatten from 'lodash/flatten'
 
 import { SearchState } from 'features/search/types'
-import { Offer } from 'libs/algolia'
 import { captureAlgoliaError } from 'libs/algolia/fetchAlgolia/AlgoliaError'
 import { buildOfferSearchParameters } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/buildOfferSearchParameters.ts'
 import { offerAttributesToRetrieve } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/offerAttributesToRetrieve'
@@ -9,6 +8,7 @@ import { client } from 'libs/algolia/fetchAlgolia/clients'
 import { buildHitsPerPage } from 'libs/algolia/fetchAlgolia/utils'
 import { env } from 'libs/environment'
 import { Position } from 'libs/geolocation'
+import { Offer } from 'shared/offer/types'
 
 type FetchMultipleOffersArgs = {
   paramsList: SearchState[]
