+++ src/libs/algolia/fetchAlgolia/fetchOfferHits.ts
@@ -1,10 +1,10 @@
 import { initialSearchState } from 'features/search/context/reducer'
-import { Offer } from 'libs/algolia'
 import { captureAlgoliaError } from 'libs/algolia/fetchAlgolia/AlgoliaError'
 import { buildOfferSearchParameters } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/buildOfferSearchParameters.ts'
 import { offerAttributesToRetrieve } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/offerAttributesToRetrieve'
 import { client } from 'libs/algolia/fetchAlgolia/clients'
 import { env } from 'libs/environment'
+import { Offer } from 'shared/offer/types'
 
 type FetchOfferHitsArgs = {
   objectIds: string[]
